Encode chart_entry to base64
This commit gets rid of the json.loads() call as the argparse type for
chart_entry in update-index. Instead we now encode the index entry to
base64 in chart-repo-manager before adding it to the GITHUB_OUTPUT and
decode it in update-index back to a dict.

This allows the use of double quotes when passing the "--chart-entry"
argument to update-index, which enables the possibility for variable
evaluation. This paves the way for using environment variables instead
of using step's output directly.

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>