Enable metadata assertion and chart lock generation
The owners module changes enables parsing Red Hat OWNERS files and asserting
that metadata meets new expectations of the content.

The packagemapping module provides hooks to generating chart locks based
on the content of a given repository's charts/ directory.

The pullrequest module changes facilitates asserting metadata given a
pull request as an input.

Entrypoints are then added to setup.cfg to simplify calling of this new
logic in CI.

Signed-off-by: Jose R. Gonzalez <komish@flutes.dev>