Initialize the Submission data structure
This is the first step towards solving #296. The code added by this
commit is currently *not* put in use in the certification pipeline and
is a partial duplicated of the "checkprcontent" package.

The Submission data structure can collect and run early validation
checks on all information concerning a user's submission (i.e. a GitHub
Pull Request). In its current state, given an api_url (link to a given
PR), it can collect the list of modified files, and extract the
category, organization, chart's name and version, as well as running
basic checks, such as SemVer compatibility of the provided version.

Eventually, this data structure is meant to encapsulate more
information concerning the user's submission, and be dumped to a file to
be passed around the certification pipeline jobs. This early (and
incomplete) version is intended to trigger discussion around the design
choices and general direction taken to solve #296.

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>