Allow ignoring the presence of an OWNERS file
When validating that a PR contains the correct set of file to certify a
chart, this PR adds the possibility to ignore the fact that an OWNERS
file is also present.

While an OWNERS file should never be present in this type of submission,
the case of an invalid Submission containing a mix of OWNERS and other
files is handled in the is_valid_owners_submission method.

Eventually, errors related to OWNERS file (which this belongs to) feed
the owners-error-message GITHUB OUTPUT, while other errors feed the
pr-content-error-message GITHUB OUTPUT.

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>