Cache virtualenv between jobs
Installation of the python dependencies and scripts is done at the start
of each job within the build.yaml workflow.

By caching the venv we can save up to 10 seconds on each job after the
first. With the planned "Submission refactoring" that splits the
chart-verifier job into 3, this can scale to up to a total of 40s per
run.

Signed-off-by: mgoerens <mgoerens@redhat.com>