Refactor get-verifiy-param to use Submission
This commit doesn't change the behavior of the get-verify-param step. It
builds upon the new Submission data structure to avoid additional calls
to the GitHub API and improve some readability aspects of the code.

This commit also introduces a slight modification to the Submission data
structure. Previously, the path related to the chart's source pointed to
the "./src/Chart.yaml" file, it now points to the "./src" directory.
This is useful as this is one of the return value of
generate_verify_options.

Finally, this commit also changes the SUBMISSION_PATH environment
variable to an absolute path. This way, steps that run from within the
"./pr-branch" directory such as get-verify-param can easily load the
submission artifact. Note that $GITHUB_WORKSPACE is the default path
under which load an artifact is downloaded by the download-artifact
action.

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>