Failover register original branch in repo_manager
Upon initialization, WorkflowRepoManager registers the repository
original branch, so that it can return to that branch during cleanup. If
the tests are run on a detached HEAD, this operation fails as it expects
to be on a git branch.

This commit adds a failover to handle that particular case, defaulting
to using the commit ID.

Signed-off-by: mgoerens <41898282+github-actions[bot]@users.noreply.github.com>