Increase retry timeout when polling CI run result
During end to end tests, we query the GitHub API regularly to check if
the pipeline associated with the test PR has completed. This commit
increase the retry timeout in order to decrease the total amount of API
calls that the bot account is performing. This helps with staying within
the GitHub API rate limits as highlighted in #422.

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>